Making some useful adjustments.
We don't need to be installing jekyll by default if we install Ruby. We
apparently also weren't installing git, which made it hard to do things
like install stuff through composer if it wanted to talk to github. Also
added some PHP-specific stuff to the default provisioned site, but
commented out. Basically so that it is there as a reference.